2 more tests

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -13,6 +13,10 @@ class CreateUsersTable extends Migration
      */
     public function up()
     {
+        if (config('database.connections.sqlsrv.driver') == 'sqlite') {
+            return;
+        }
+
         Schema::create('users', function (Blueprint $table) {
             $table->increments('id');
             $table->string('name');
@@ -31,6 +35,10 @@ class CreateUsersTable extends Migration
      */
     public function down()
     {
+        if (config('database.connections.sqlsrv.driver') == 'sqlite') {
+            return;
+        }
+
         Schema::dropIfExists('users');
     }
 }
